Quick hack on Stahl19/20
